refactor(product page): Rename main product section component
