+++ frontend/templates/product/sections/ProductOverviewSection/AddToCart/AddToCartBar.tsx
