+++ frontend/templates/product/sections/ProductOverviewSection/CompatibleDevices.tsx
