+++ frontend/templates/product/sections/ProductOverviewSection/ProductDescription.tsx
