+++ frontend/templates/product/sections/ProductOverviewSection/ProductGallery.tsx
