+++ frontend/templates/product/sections/ProductOverviewSection/ProductRating.tsx
