+++ frontend/tests/jest/tests/ProductReviews.test.tsx
@@ -6,7 +6,7 @@ import {
    getReviewsResponse,
 } from '../utils';
 import { ReviewsSection } from '@templates/product/sections/ReviewsSection/index';
-import { ProductRating } from '@templates/product/sections/ProductSection/ProductRating';
+import { ProductRating } from '@templates/product/sections/ProductOverviewSection/ProductRating';
 import { useProductReviews } from 'templates/product/hooks/useProductReviews';
 
 jest.mock('@templates/product/hooks/useProductReviews', () => ({
